Moved language to AppearancePreferences
